only delete when there is sth to delete

+++ controllers/PurchaseController.php
@@ -6,12 +6,14 @@ class PurchaseController extends ShoppingControl_Controller_Action
     public function deleteAction()
     {
         $purchaseIds = $this->_getParam('purchase_id');
-        $purchaseIdsConcat = implode(',', $purchaseIds);
-        $db = Zend_Registry::get('db');
-        $db->delete(
-            'purchase',
-            sprintf('purchase_id IN (%s)', $purchaseIdsConcat)
-        );
+        if (!empty($purchaseIds)) {
+            $purchaseIdsConcat = implode(',', $purchaseIds);
+            $db = Zend_Registry::get('db');
+            $db->delete(
+                'purchase',
+                sprintf('purchase_id IN (%s)', $purchaseIdsConcat)
+            );
+        }
         // Redirect us back to where we came from
         $this->_redirect($_SERVER['HTTP_REFERER']);
     }
